Restructure the test layout
Instead of the *_spec_helper files we only provide:

* no_rails_spec_helper (for tests that don't need rails)
* rails_spec_helper (for tests that need rails)

Also introduce spec/controllers, so rspec/rails is happy about where contorllers
live. Now we sort of need to be strict about the type of example groups in which
we include helpers. This means that when you define a helper file in spec/support
you should do:

    module Blah
      # ... helpers
    end

    RSpec.configure do |config|
      config.include Blah, type: <THIS IS THE IMPORTANT BIT>
    end